Added CLI.PY and LOGIC.PY
Improved CLI structure: moved argument parsing and logic(handling) into separate modules

Main.py now just runs main; logic has been moved into cli.py and logic.py

Print feedback should still be added for each node for better user experience
